+++ CHANGES.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- Read functionality for film chunks.
+
+### Changed
+
 - Remove ability to pass `None` as an argument for `requests_per_second` when creating a client.
 
 ## [0.8.0] - 2024-05-11